Fix event filtering.
The fix had to be made because the program was filtering PR events just
because the subject was specified in the user configuration and the score
matching was horrible.

That's why in the implementation I used an approach where I figured out
the general groups (like for PR) and use that list to remove general
groups from an events groups - which allowed me to better compare groups
specified by the user.